Update and rename server.js to app.js
